Remove laxMode
Unfortunately, vue-router overrides window.history.back
and window.onpopstate with its own stuff. So the modals
would close in lax mode, but the browser history stayed
polluted with the now-inaccessible modal history
entries.

I considered making Flows restore the modal that had
been navigated away from if the user tried to return to
it, but this could not be done because vue-router
overrides Flows' onpopstate listener when the route
changes.

It seems to me that the only non-buggy behavior is to
completely reject route changes from inside modal
flows.