Set up rollup build script
I had to move the demo project into a subfolder so it
wouldn't be included in the dist root folder. It's really
strange. I think there's a better way to do it. Also, I
have yet to get the import working in another project.

Really not a fan of how complex and finnicky typescript
makes everything.