Debug build process
It finally works! It outputs all the correct files, with
their type definitions! Unfortunately, the Vue extension
declaration file is still not included, but I'm able to
use the library in another project with no errors other
than that one.

It also no longer includes the demo files in the built
package output.